sources: do not override datasource detection if None is in list
Users with datasource_list = [Azure, None] started failing to boot
properly outside of Azure with the changes to override datasource detection.

If the fallback "None" is included in the datasource_list, do not treat
the system as configured with a single datasource.

If users want to force a single datasource regardless of detection,
they can do so by removing None from the list.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>