tools/ds-identify: match Azure datasource's ds_detect() behavior
The current identification enables the Azure datasource for all
of Hyper-V VMs.  This datasource is strictly intendeded to work
on Azure cloud as identified by DMI chassis tag and has largely
been limited to this until 23.2 with a guard in the datasource.

Match ds_detect() behavior by removing hypervisor and filesystem
label checks.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>