Add cloud-init overview/introduction
One thing cloud-init has been lacking is a high-level overview of
what it is, what it does, and how it works. Once we are happy with
the general structure and content of this introduction, I will also
add links to the various pages around the docs that a newcomer might
want to visit for more information on each of the topics.