net: allow for NetworkManager configuration for NetworkManager net renderer
This is the equivalent of the change made in the following commit for sysconfig
renderer but this time it is for network_manager renderer:
67bab5bb804e2346673 ("net: Allow for NetworkManager configuration")

The reasons for making the change has already been outlined in the above commit.
When cloud-init configures DNS and produces /etc/resolv.conf, NetworkManager,
if enabled, can clobber the above file and could also render it empty.
Hence, this change makes cloud-init add a NetworkManager configuration so that
it disables management and handling of /etc/resolv.conf by NetworkManager.

Some code refactoring has been performed so as to avoid code duplication.
Unit tests have been adjusted accordingly.

LP: #1693251

Signed-off-by: Ani Sinha <anisinha@redhat.com>