Ensure known package managers can be used
Previously, a package manager would only be "known" if it was initialized,
but this won't happen if it isn't a default for the distro.
E.g., Debian never initializes Snap, so it is never seen as a known package manager.
Instead, be explicit about which package managers are known.
I moved it into a util file to avoid a circular import.

The removed code in `install_packages` (lines 225-231) were redundant
because we already make
that check in `_extract_package_by_manager`