sources/Azure: Ignore system volume information folder while scanning for files in the ntfs resource disk (#4446)
In VM SKUs with large ephemeral disk, the NTFS-formatted resource disk
might have the SVI (System Volume Information) folder present. This 
causes cloud-init to skip formatting the resource disk for these SKUs 
during VM stop/start. This PR adds the SVI folder to the list of files 
cloud-init should ignore when it attempts to check if there is any user 
data on the resource disk.