azure/imds: retry on 429 errors for reprovisiondata
Allow retries for 429 when fetching reprovisiondata. 429 is used when
exceeding rate-limits and is currently retried for fetching instance
metadata.

A tight retry-loop was observed on older versions of cloud-init where
we had a series of short (2 second) read timeouts.  We expect this had
triggered the rate-limiting due to max concurrent connections (from
IMDS' perspective). Then _poll_imds() then entered a tight retry loop,
spamming the logs until connections were allowed to resume.

The standard sleep between readurl() retries should be an adequate
backoff if 429 is observed.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>