cmd: Don't write json status files for non-boot stages (#4478)
Don't use status_wrapper() for `cloud-init modules --mode init`.
      
Both status.json and result.json are currently written when `cloud-init
modules --mode init` is ran. This is unecessary since this is not a boot
command used by cloud-init, and as a workaround for previous bugs related to
this behavior this leaves behind unwanted keys in status.json.

This aligns cloud-init code's behavior with documented behavior.