summit-notes: add 2023 notes for reference in mailinglist/discourse
Provide an unauthenticated view of meeting notes for those without
google accounts.  Original meeting notes live in a shared google
drive document. To provide a point of reference for those in the
community without google accounts we have refactored the notes to
markdown and will host them in the cloud-init repository so that
both discourse trip reports and mailinglist threads can reference
these notes without have to bloat the email or discourse post with
details.

Co-authored-by: Sally Makin <sally.makin@canonical.com>