refactor(subp): Remove parameter 'combine_capture' (#4555)
This parameter has a single callsite. Configuration modules shouldn't
send output directly to stdout / stderr.

Note: Regarding the code below the comment "Just ensure blank instead of none"

'None' output was only possible for the stderr tuple member output of
.communicate() when stderr was directed to STDOUT. This only occurred in the
case of the 'combine_capture' argument, and only for stderr, not stdout. This
removes the unnecessary assignments for the 'capture' argument, as well
as the no-longer-necessary assignments for 'capture_parameter'.

See the docs[1] for more details.

[1] https://docs.python.org/3/library/subprocess.html