netplan: Wait for dbus.socket (#4592)
cloud-init.service is ordered after systemd-networkd-network-wait-online.service,
which is ordered after systemd-netword.service. Therefore, during
cloud-init.service, systemd-networkd has started and is running, but is not
listening on dbus if dbus is not yet available. This results in failure running
`netplan apply`. If this occurs, retry until dbus.socket is available.