fix: Move cloud-final.service after time-sync.target
The cloud-final service (which may include apt or HTTPS operations
requiring certificate verification) must run after the clock has been
synchronized. Otherwise, systems lacking an RTC (notably the Raspberry
Pi) risk hitting the certificates that are valid "in the future" because
their clock is wrong.

This change was tested on Ubuntu mantic (23.10) server images on a
Raspberry Pi both connected to a LAN, and disconnected from a LAN. In
both cases, start-up time was not affected significantly, and
cloud-final started after time-sync had finished (successfully or
otherwise).

Fixes #3928
LP: #1951639