azure: Check for stale pps data from IMDS (#4596)
After reprovisioning, IMDS should return None value for PPS data
type.

Introduce a check to verify it IMDS is returning stale-PPS
data. In case of stale data, report the error to the host and allow
the VM to continue with provisioning. This is part of a larger effort
to improve error handling and error reporting during VM provisioning.