test: Remove snapshot option from install_new_cloud_init (#4682)
Integration tests allow for installing a new version of cloud-init with
the option of snapshotting the image after install. After taking
a snapshot, it will then set the snapshot as the default for *all*
future launches in this test run. This is also used by the conftest to
setup the initial image, but it should not be used anywhere else.

Rather than having it as an option to the install_new_cloud_init method,
remove the option and make the confest set the global snapshot instead.
This removes a footgun when writing other tests that need to install a
different version of cloud-init.