fix(ds-identify): relocate /run on *BSD (#4677)
shuffle code-paths around such that we initialize RUN_PATH, and all the
things that depend on it after figuring out which OS we're on, while
still alowing these global variables to be overwritten from the outside.
On BSD /run is not ephemeral, so here we set RUN_PATH to /var/run.
Ensure everything is initialized for 'print_info', not just for 'main'.

For the tests, we equally ensure that for not-Linux systems the data is
written into `rootd + var/run` instead of `rootd + run`.

Sponsored by: The FreeBSD Foundation