test: Fix silent swallowing of unexpected subp error (#4702)
We throw an error when supb isn't mocked in a test. However, there
are many broad exception handlers that wind up swallowing the error.
This commit changes it from an AssertionError to a new error that
inherits from BaseException so that it doesn't get silently swallowed.

Also fix any tests that are impacted.