logging: Add log message when create_hostname_file key is false
Setting create_hostname_file to False while trying to set the
hostname via userdata can result in the hostname not being
correctly set (it is overwritten by the cloud provider's DHCP
service).  This change adds a log line every time a hostname file
is not created due to create_hostname_file being False for easier
debugging of that situation.