Fix Alpine and Mariner /etc/hosts templates
Revert the changes I introduced with #1084, these had the side-effect
of causing "hostname" problems. Fixes #4771.

Also remove the "::1 {{fqdn}} {{hostname}}" entry as this this being
results in strange behaviour with both Busybox "hostname -i" (returns
"127.0.0.1", or "127.0.0.1 A.B.C.D" if an additional non-loopback IP
entry in /etc/hosts) and Coreutils "hostname -i" (returns "::1" in
both situations), commands whereas if the "::1" entry is removed then
both Busybox and Coreutils behave the the same fashion returning
"127.0.0.1" or "127.0.0.1 A.B.C.D" as appropriate.

Also correct hosts.mariner.tmpli file it was not valid (every line
was prefixed with "+").