Modify ds-identify for Alpine support and add OpenRC init.d script
Add OpenRC init.d script for ds-identify.

Add support to ds-identify for "virt-what" as an alternative to
"systemd-detect-virt" for non-systemd Linux distros.

Change ds-identify to use short option for "modprobe" for Busybox
compatibility.

Change ds-identify to use short options for "grep" as POSIX does not
define long options, they are a GNU-ism.