refactor: Remove unused networking code
distros/__init__.py employs two ways of configuring network.
There was an old way in which the distro directly wrote out its
networking config, and a new (as of 7 years ago) way which involves
selecting a network renderer and rendering network using that.

We have kept the old way around to support Arch and Gentoo, whose code
never switched to using the new renderers, and write netctl and netifrc
respectively. This code is no longer needed.

On Arch, the only networking system included in the base image is
systemd-networkd, and the only supported init system is systemd. The
networkd renderer automatically gets selected in these images, so the
netctl code is no longer needed.

On Gentoo, the systemd images act the same as Arch. Due to a bug in
cloud-init, the openrc images never fall back to the netifrc code,
and instead throw an exception when they cannot find a suitable
renderer. This behavior has been in place for several years, which
means that the netifrc code has gone unused for several years.

If there's value in either of these forms of writing network, we can
always bring them back as separate renderers, but removing them
simplifies the distro code immensely.