fix(source): fix argument boundaries when parsing cmdline
When the command-line includes an argument that contains `ds=`, the rest
of the argument is always interpreted as a datasource, regardless of
wether the argument is `ds` or just the suffix of another argument.
While this doesn't cause issues in a majority of cases, it starts
conflicting with other arguments that also include `ds=` but do not
refer to a cloud-init datasource. For instance, setting
`bonding.max_bonds=0` in the command-line (which is a real-world
use-case) conflicts with the current implementation.

To fix this issue, this commit ensures that `ds`, `ci.ds` and
`ci.datasource` arguments are either at the very beginning of the
command-line, or just before a whitespace, so that we don't match
another argument.