distro: add eject FreeBSD code path
OpenBSD, and NetBSD both have an eject(1), so they should be covered in
the default code path.

FreeBSD and Dragonfly however, do not have eject in base. Here, eject is
an (unmaintained) port. In base, we do however, have camcontrol(8) and
cdcontrol(1), both of which have an eject subcommand.

Let's use camcontrol(8) here.

Sponsored by: The FreeBSD Foundation