cc_growpart.py: fix handling of /dev/mapper devices for Alpine
Unlike the majority of other Linux distros /dev/mapper entries are NOT
symlinks to /dev/dm-X block devices but rather are actual block devices
themselves. This is a side-effect of Alpine's use of mdev, rather than
udev, in its initramfs.

Therefore this change works out the corresponding /dev/dm-X for any
/dev/mapper/XYZ device on Alpine.