fix(hotplug): remove literal quotes in args
Remove literal double quotes which cause
/usr/bin/cloud-init devel hotplug-hook: error: argument -s/--subsystem: invalid choice: '"net"' (choose from 'net')
in cloud-init-hotplugd.service when a NIC is added/removed.

Literal quoutes provoke errors, an eval or something in those lines
would bypass the error, but that would be equivalent to no quotes at
all.

This quouting was introuced in e70cd47846120b83035daa87e089fa045355f03a.

Fixes: GH-4886