test(ubuntu-pro): change livepatch to esm-infra
Livepatch is reported as warning by ubuntu-pro-client when enabled on
some unsupported kernels.

The test is only concern with the fact that only custom services defined
in user-data are enabled by cloud-init / ubuntu-pro-client.

Change livepatch to esm-infra, as enabling this service should be much
more stable.

Consider future warnings as ok, for what this test is concerned.

Co-authored-by: Chad Smith <chad.smith@canonical.com>