test: fix verify_clean_log
verify_clean_log assumed shadowed warning messages to be contained in a
single line entry with a `[WARNING]`, but warnings can be multiline as:

```
2024-02-16 15:01:49,354 - activators.py[WARNING]: Running ['netplan', 'apply'] resulted in stderr output: Cannot call Open vSwitch: ovsdb-server.service is not running.
Failed to connect system bus: No such file or directory
Falling back to a hard restart of systemd-networkd.service
```

wrongly asserting the number of expected_warnings.

Make this check more reliable by asserting that the number of lines
containing `WARNINGS` is less or equal than the number of appearances of
the shadowed warning messages.