fix(ds-identify): Prevent various false positives and false negatives (#4905)
These issues represent a class of error that is difficult to reason about
for users and developers alike because when parsing fails with a broken
parser it isn't obvious whether it is a user error or a cloud-init bug. Source
inspection and trial-and-error testing are a users' options to resolve this
issue, and neither option is a good option.

Using grep in check_config() causes 2 classes of false positives:

- Matches in comments aren't ignored.
- Callsite that require the key to have a value are broken.

Using grep in check_config() causes 2 classes of false negatives:

- Keys are assumed to not be single quoted.
- Valid whitespace breaks parsing in various ways.

The fix added two new functions:

get_value() - for a k/v pair check that the value is non-empty
get_single_line_flow_sequence() - for a k/v pair check that the value is
    a valid single line flow sequence [1]

Add various tests for valid yaml.

[1] https://yaml.org/spec/1.2.2/#741-flow-sequences