networkd: Always respect accept-ra if set
Support the explicit rejection of router advertisements in cases where
the network may be configured to send them, but they are not desireable,
for example, in a case where static IPv4 addressing is used.

Before this change, IPv6AcceptRA is only included in the rendered
networkd files if DHCP is used.

Now, if a boolean value for accept-ra is present, it will be included in
the rendered output, regardless of the configured addressing.

Fixes: #4927