d/control: Add build-depends on pybuild-plugin-pyproject
This switches the pybuild backend plugin from distutils to pyproject.
Distutils, the most commonly used build system plugin, is deprecated[1]
and removed from the standard library in upstream Python starting in
3.12. This plugin still works on Noble (which provides Python 3.12) due
to the python3-distutils package, which is based on distutils from Python
3.11.

The pyproject plugin is expected to replace distutils, due to its
support of PEP-517. Upstream cloud-init already supports PEP-517[2].

To enable building with the pyproject build system,
pybuild-plugin-pyproject must be a build dependency, as well as any
build tools specified in pyproject.toml[3]. Cloud-init already build
depends on python3-setuptools, so only pybuild-plugin-pyproject needs
to be added.

The pybuild-plugin-pyproject package is in universe, however it is
acceptable because this is a build-only dependency[4].

[1] https://docs.python.org/3.10/whatsnew/3.10.html#distutils-deprecated
[2] https://github.com/canonical/cloud-init/commit/8836c7824533025768ec82c350a965b76a744b1e
[3] https://manpages.ubuntu.com/manpages/noble/man1/pybuild.1.html
[4] https://github.com/canonical/ubuntu-mir (see [Dependencies] section)