Cleanup test_net.py (#4840)
test: cleanup test_net.py

* Move ~3500 lines of network config into its own file
* Replace unittest idioms with pytest idioms
* Remove CiTestCase and FilesystemMockingTestCase
* Parametrize the obvious tests
* Remove commented tests and debug print statements

Much of the heavy lifting was done using `pytestify`.
Moving configuration into another file was mainly because the file
is way too large.

We have commented tests that are years old. Git will keep the
history if we ever need to bring them back. I left the TODO
in place as that can still be a reminder that they are needed.