chore(ephemeral): add debug log when bringing up ephemeral network (#5010)
A debug message is logged whenever an instance is able to connect 
to the IMDS but when it can't, and an ephemeral network needs 
brought up, nothing is logged. Adding a log message here will help 
improve readability of logs and help troubleshoot failure paths 
when an instance is having issues with networking.