test: Fix scaleway retry assumptions
In test_scaleway.py, test_metadata_connection_errors_legacy_ipv4_url
assumes that a call to wait_for_url() will retry the number of
retries specified by the datasource, but that retry number is used
for readurl() calls, not wait_for_url() calls. The fact that the
assertion happened to work previously was entirely incidental.

Additionally, mock time.time() and time.sleep() calls so that we're
not needlessly waiting in unit tests.