feat(hotlug): trigger hotplug after cloud-init.service
Extend hook-hotplug to trigger cloud-init-hotplugd.socket after
init-local and init boot stages have been completed by checking
status.json.

Previously, it was only triggered after the cloud-init was completed,
leaving a huge gap where hotplugged NICs weren't processed.

The reason to not extend the scope of this to after init-local is
completed is that during init, datasources are fetched and pickled and
hotplugd.service interacts with the ds, thus further syncing has to be
implemented to achieve that.

Fixes: GH-4863