feat(hotlug): trigger hotplug after cloud-init.service (#5058)
Extend hook-hotplug/cloud-init-hotplugd.socket after
the init boot stage has been completed on systemd
systems to send but keep restricting cloud-init-hotplugd.service after
cloud-init has finished, to avoid ds pickling conflicts.

Previously, it was only triggered after the cloud-init was completed,
leaving a huge gap where hotplugged NICs weren't processed.

The reason to not extend the scope of this to after init-local is
completed is that during init, network config can be applied if
BOOT_LEGACY is configured as an update event and could race with hotplug.

Fixes: GH-4863