bug(maas): register the correct DatasourceMAASLocal in init-local
Incorrectly registering DataSourceMAAS for init-local DEP_FILESYSTEM
inadvertently short-circuits the _get_data checks performed by
DataSourceMAASLocal for presence of network config from initramfs.
This results in a 2 minute timeout on physical system
reboots because the network stage DataSourceMAAS expects all system
networking is already configured and the IMDS to bebe accessible.

Correct this registration to do the proper initramfs checks and avoid
trying to perform GETs against the IMDS in init-local timeframe when
we know network is not yet available.

Fixes GH-5064