Avoid network configuration if we already generated a udev file
In a case where the initialization of an instance gets interrupted by a reboot
it is possible that we have already generated the udev rules for the
network devices. When the system comes up again we are in a race with udev
to apply any renaming that may need to be done. If udev gets there first
the device will be in a down state after cloud-init is finished. Set the
semaphore to indicate the network is already configured if we find the
generated udev rules file.