Parse and load non-YAML config files
Previously, configuration files were assumed to be YAML only. This
change allows for non-YAML files to be loaded by just setting
userdata_raw to a list of binary data instead of attempting to parse
each config manually. This has the consequence of not being able
to control how merging works for those files.