fix: Ensure network config in DataSourceOracle can be unpickled (#5073)
`_network_config` is only explicitly set on the instance if
`_is_iscsi_root()` is True. This means that when `_is_iscsi_root()` is
False, we're modifying the class variable directly. When the instance
gets pickled, the class variable is not included so such changes do
not get persisted. This commit fixes this.