util: add log_level kwarg for logexc()
Many failures are being treated as warnings instead of errors due
to usage of logexc() to emit the failure.

Add log_level parameter to allow increasing the log level without
requiring an additional log.

Add tests but I'm unsure why its not logging the backtrace when
the failure occurs within the test method.

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>