fix(url_helper): fix TCP connection leak on readurl() retries (#5144)
In scenarios where a lot of retries are expected, Ubuntu 24.04 fails
regularly with "Too many open files".  The `ulimit -n` shows the same
number of allowed open files in Ubuntu 20.04 (1024), but the connections
don't close on 24.04. As retries gets close to 1024 in readurl(), the
open file limit is hit and exceptions sprout up in a number of places.

It appears that the reuse of Sesssion's context manager triggers a
connection leak on python3-requests used in 24.04 when saving references
to the requests (in excps[]).

- drop `with session as sess` context manager. Session should be able
  to handle all retry attempts without a context manager

- raise exceptions immediately when required rather than saving them to
  excps[] to raise outside of the exception handler

Signed-off-by: Chris Patterson <cpatterson@microsoft.com>