fix: Warn when signal is handled
Problems:

- When systemd stops cloud-init, no warning is logged.
- When systemd restarts cloud-init, status info and return code from the
  current stage is lost.

The net effect of this is that when cloud-init is restarted, unless one
explicitly checks the logs for the signal message, no warning is logged,
and cloud-init status indicates no problem.

Communicate unexpected state by:

- Log an error when handling signals.
- When sys.exit() is called from a signal, write status.json before exiting.
- Append errors and recoverable errors rather than overwriting them.
- Warn when status.json contains a time when the stage is starting[1]

This should make cloud-init properly warn in status output and logs that
something is amiss.

[1] This is a bit cautious, however when cloud-init doesn't have the
    opportunity to handle a signal (SIGKILL), the other mitigations will fail.