fix: Ensure dump timestamps parsed as UTC
When parsing timestamps in `cloud-init analyze dump`, most log parsing
used UTC, but if shelling out to `date` was required, it used the local
timezone. This is wrong because cloud-init logs are in UTC.
This commit ensures that all timestamps are UTC-based.

Also update test code to explicitly convert generated times to UTC
before checking against what is being tested.

Fixes GH-5158