Fix the logic of handling /usr/etc/sudoers
There is no need to copy the content of the file. If we do not find the
include directive in either the /etc/sudoers file or the /usr/etc/sudoers
file we append the include directive to the /etc/sudoers file if it exists,
if it does not exist we simply write the include directive into /etc/sudoers.
This preserves the current behavior w.r.t. oredering of the read operation.