device_part_info: do not recurse if we did not match anything (#5169)
First off, the recursion was bogus, because it called the wrong
function, but more importantly, it would lead to infinite recursion
anyway, even if we called the right one.

Throw a ValueError instead, because we've most likely been called with a
bogus value. Add a test to show that.

Sponsored by: The FreeBSD Foundation