bug(package_update): avoid snap refresh in images without snap command
When package_update or package_upgrade: true are provided in
cloud-config do not call snap refresh on systems that do not have
the snap command.

This was intended as fixed in cbe5f3a119 but the fix only avoided
snap refresh during package install, not the generic package update
operation.

Fixes: GH-5143
LP: #2064132