fix(cli): retain support for --files argument before subcommand
Upstream dropped support for a generic positional --file argument ordered
on the command line before the cloud-init subcommand because the
--files argument is only applicable to the specific subcommands init,
modules and single.

Add functionality to retain support for --file or -f argument both
before and after the subcommand to avoid regressions in stable releases due
to a change in behavior.

LP: #2064300